chore(resources): remove the legacy gitlab://schema resources
RegisterDynamicSchemaResources and RegisterMetaSchemaResources served
the first-generation schema-discovery URIs (gitlab://schema/dynamic/*,
gitlab://schema/meta/*). The surface-aware tool manifest (gitlab://tools
+ gitlab://tools/{id}) replaced them and cmd/server stopped registering
them; since then their only callers were their own unit tests, and the
'survive for isolated tests and audits' note in doc.go/cachehints was
stale — no audit or e2e ever read them. Per the project's strict-spec
posture, abandoned surface gets removed, not curated.

What actually survives moved to tool_manifest.go, because the served
manifest uses it: the required-params extraction helpers and the schema
enrichment chain (guidance + destructive markers) that gitlab://tools/
{id} details are built from. Their tests moved along, with the
enrichment integration test rewired to read through the served detail
instead of the deleted URI — proving the enrichment reaches the live
surface, which the old test never did. Net: -1,095 lines of dead
surface, zero behavior change on anything a client can reach.
