+++ README.md
@@ -423,20 +423,20 @@ and is never logged. Full details: [PRIVACY.md](PRIVACY.md).
 
 | Category                 |     Files |       Lines |
 | ------------------------ | --------: | ----------: |
-| Source (`.go`, non-test) |       993 |     201,251 |
-| Unit tests (`_test.go`)  |       550 |     311,091 |
+| Source (`.go`, non-test) |       991 |     200,959 |
+| Unit tests (`_test.go`)  |       548 |     310,669 |
 | End-to-end tests         |       174 |      45,163 |
-| **Total**                | **1,717** | **557,505** |
+| **Total**                | **1,713** | **556,791** |
 
 ### Functions
 
 | Category                        |  Count |
 | ------------------------------- | -----: |
-| Source functions                |  7,508 |
-| — exported (public)             |  2,617 |
-| — unexported (private)          |  4,891 |
-| Unit test functions (`TestXxx`) | 11,675 |
-| Subtests (`t.Run(...)`)         |  2,934 |
+| Source functions                |  7,496 |
+| — exported (public)             |  2,615 |
+| — unexported (private)          |  4,881 |
+| Unit test functions (`TestXxx`) | 11,662 |
+| Subtests (`t.Run(...)`)         |  2,931 |
 | End-to-end test functions       |    381 |
 
 ### Ratios worth noting
@@ -445,17 +445,17 @@ and is never logged. Full details: [PRIVACY.md](PRIVACY.md).
 | ---------------------------------- | -------------------------: |
 | Test lines vs source lines         | 1.55× more tests than code |
 | Average source file length         |                 ~202 lines |
-| Average test file length           |                 ~565 lines |
-| Comment lines in source            |  23,149 (~11.5% of source) |
+| Average test file length           |                 ~566 lines |
+| Comment lines in source            |  23,074 (~11.5% of source) |
 | Test functions per source function |                       1.6× |
 
 ### Code patterns
 
 | Pattern                            | Count |
 | ---------------------------------- | ----: |
-| `if err != nil` checks             | 6,696 |
+| `if err != nil` checks             | 6,682 |
 | `defer` statements                 |   911 |
-| `struct` types defined             | 2,727 |
+| `struct` types defined             | 2,723 |
 | `//nolint` suppressions            |   258 |
 | `TODO` / `FIXME` / `HACK` comments |     2 |
 
@@ -478,8 +478,8 @@ and is never logged. Full details: [PRIVACY.md](PRIVACY.md).
 
 | Fact                                 | Value                                                                                                |
 | ------------------------------------ | ---------------------------------------------------------------------------------------------------- |
-| Source code printed at 55 lines/page | ~3,659 pages of A4                                                                                   |
-| Source lines mentioning `"gitlab"`   | 12,593 (impossible to avoid)                                                                         |
+| Source code printed at 55 lines/page | ~3,653 pages of A4                                                                                   |
+| Source lines mentioning `"gitlab"`   | 12,572 (impossible to avoid)                                                                         |
 | Longest function name in source      | `baseDestructiveEarlySinglePromptTemplateAndFixtures` (51 chars)                                     |
 | Longest test function name           | `TestRequiredMissingAndUnknownParamNames_SchemaValidation_ReturnsSortedMissingAndUnknown` (87 chars) |
